# docs/swagger/swagger.yaml
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://github.com/erycamel/go-gin-example/blob/master/LICENSE
  termsOfService: https://github.com/EDDYCJY/go-gin-example
  title: Golang Gin API
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: TagID
        in: query
        name: tag_id
        type: integer
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[{"id":3,"created_on":1516937037,"modified_on":0,"tag_id":11,"tag":{"id":11,"created_on":1516851591,"modified_on":0,"name":"312321","created_by":"4555","modified_by":"","state":1},"content":"5555","created_by":"2412","modified_by":"","state":1}],"msg":"ok"}'
          schema:
            type: string
      summary: 获取多个文章
    post:
      parameters:
      - description: TagID
        in: query
        name: tag_id
        required: true
        type: integer
      - description: Title
        in: query
        name: title
        required: true
        type: string
      - description: Desc
        in: query
        name: desc
        required: true
        type: string
      - description: Content
        in: query
        name: content
        required: true
        type: string
      - description: CreatedBy
        in: query
        name: created_by
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章
  /api/v1/articles/{id}:
    delete:
      parameters:
      - {}
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":400,"data":{},"msg":"请求参数错误"}'
          schema:
            type: string
      summary: 删除文章
    get:
      parameters:
      - {}
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"id":3,"created_on":1516937037,"modified_on":0,"tag_id":11,"tag":{"id":11,"created_on":1516851591,"modified_on":0,"name":"312321","created_by":"4555","modified_by":"","state":1},"content":"5555","created_by":"2412","modified_by":"","state":1},"msg":"ok"}'
          schema:
            type: string
      summary: 获取单个文章
    put:
      parameters:
      - {}
      - description: TagID
        in: query
        name: tag_id
        type: string
      - description: Title
        in: query
        name: title
        type: string
      - description: Desc
        in: query
        name: desc
        type: string
      - description: Content
        in: query
        name: content
        type: string
      - description: ModifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":400,"data":{},"msg":"请求参数错误"}'
          schema:
            type: string
      summary: 修改文章
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"lists":[{"id":3,"created_on":1516849721,"modified_on":0,"name":"3333","created_by":"4555","modified_by":"","state":0}],"total":29},"msg":"ok"}'
          schema:
            type: string
      summary: 获取多个文章标签
    post:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章标签
  /api/v1/tags/{id}:
    delete:
      parameters:
      - {}
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除文章标签
    put:
      parameters:
      - {}
      - description: ID
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: ModifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 修改文章标签
swagger: "2.0"